QGlobalStatic: suppress -Wtsan warning
QGS employs a call to std::atomic_thread_fence. Unfortunately TSAN does
not support it, and GCC >= 11 raises a warning. This breaks the build
when building under -Werror. Suppress the warning using the usual
pragmas.

There's a catch: qglobalstatic.h is built into a PCH, and GCC <= 13
will still generate a warning because of
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53431 .

Change-Id: I770f39b7563b66f483851444cd580bcafc5f288a
Pick-to: 6.7 6.6 6.5
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>