QTest::qWaitFor: scope a variable tighter
When fa296ee1dcf4c16fb6f242cf08949485e499fec3 ported this function
from int timeout to QDeadlineTimer, the need to keep this variable
outside the do-while so it could be checked in the loop exit condition
fell away.

Moving the definition of the variable to the first (and only)
assignment makes the code clearer and the variable a constant.

Amends fa296ee1dcf4c16fb6f242cf08949485e499fec3.

Pick-to: 6.7
Change-Id: I7a0fe01dc68ff140beeb0e76b141c84d4bd28458
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>