dwrite: Fix crash when loading unsupported font in QRawFont
When we tried loading an unsupported font, there was an unchecked
nullptr which caused crashes later.

This is a hotfix for 6.6 and older versions, as in Qt 6.7 it is
already part of 500be123f4fbb60593fe25d5fe70fbc62445e8aa.

Pick-to: 6.2 5.15
Fixes: QTBUG-123454
Change-Id: I954a39b5e26714b1cc197d2870e9caef87b0423e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 282f30ab09ff11d977c5978934d450ea9aded036)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>