Fix end-of-parse fixup of two-digit year in QDateTimeParser
When a date string contains day of the week, day of the month, month
and two-digit year, it was still possible to get a conflicting result
in the default century instead of a consistent result in the next (in
fact present) century.

The actual logic needed to get the right year has to take into account
all date fields. This is all worked out already in actualDate(), so
delegate to it instead of trying to make do with just the year info.

Pick to 6.5 required removal of the base year argument being passed to
actualDate(). The lack of a default century for two-digit years means
the century-fixing is cruder, only trying the next and previous, but
this should suffice for the most common case.

Fixes: QTBUG-123579
Change-Id: Id057128d8a0af9f3a7708d0ee173f854bb1a2a8e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit 3e426182e2e8122d96b208702faaf3177f3a3081)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 23b69855c0d2ed747a73b15f72f1788804bdb477)
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>